allocate in different components
