+++ calls/call.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
+    <link rel="stylesheet" type="text/css" href="./call.css" />
+  </head>
+  <body>
+    <div>
+      <button class="logCall">Log Call</button>
+      <button class="phoneCall">Make a phone call</button>
+      <p>
+        Call a contact from this record. Or log a call activity to keep track of
+        your discussion and notes. <a href="">Learn more</a>
+      </p>
+    </div>
+
+    <div class="popWindow"></div>
+  </body>
+</html>
