+++ signUp/index.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Sign up</title>
+    <link rel="stylesheet" href="./style.css" type="text/css" />
+  </head>
+  <body>
+    <img class="logo" src="../src/img/byteCRM.png" alt="" />
+    <div class="formContainer">
+      <form class="signUpForm" action="">
+        <p class="asking">
+          Have an account?&nbsp;<a class="signInLink" href="">Sign in</a>
+        </p>
+
+        <h3>Create your free account</h3>
+        <span class="slogan">Free forever. No credit card needed</span>
+        <div class="googleSignUpWrapper">
+          <img class="googleIcon" src="../src/img/googleIcon.svg" alt="" />
+          <button class="googleSignUpBtn">Sign up with google</button>
+        </div>
+        <hr />
+        <label class="inputLabel" for="">First Name</label>
+        <br />
+        <input class="firstNameInput" type="text" />
+        <br />
+        <label class="inputLabel" for="">Last Name</label>
+        <br />
+        <input class="lastNameInput" type="text" />
+        <br />
+        <label class="inputLabel" for="">Email address</label>
+        <br />
+        <input class="emailInput" type="email" />
+        <br />
+        <label class="inputLabel" for="">Password</label>
+        <br />
+        <input class="pwdInput" type="password" />
+        <br />
+        <button class="submitBtn" type="submit">Sign Up</button>
+      </form>
+    </div>
+  </body>
+</html>
