+++ signUp/successful.css
@@ -1,0 +1,37 @@
+@import url('https://fonts.googleapis.com/css2?family=Varela+Round&display=swap');
+
+body,
+h1,
+h3 {
+  margin: 0;
+  padding: 0;
+  font-family: 'Varela Round', sans-serif;
+  background-color: #f5f8fa;
+}
+
+div {
+  margin-left: 35%;
+  margin-top: 10%;
+}
+
+img {
+  width: 20rem;
+  height: 20rem;
+}
+
+h1 {
+  color: #0091ae;
+}
+
+h3 {
+  color: #0091ae;
+  margin-left: 10%;
+}
+
+a {
+  color: #0091ae;
+}
+
+a:hover {
+  color: #33475b;
+}
