+++ signUp/successful.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" type="text/css" href="./successful.css" />
+    <title>Successful</title>
+  </head>
+  <body>
+    <div>
+      <h1>Successfully Sign Up!</h1>
+      <img src="../src/img/tick.svg" alt="" />
+      <h3>Back to <a href="">Login</a> ?</h3>
+    </div>
+  </body>
+</html>
