+++ src/pages/LoginForm/components/logo.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Logo from '../../../img/byteCRM.png';
+
+const CRMLogo = () => {
+  return (
+    <div>
+      <img className="byteCRMLogo" src={Logo} alt="" />
+    </div>
+  );
+};
+
+export default CRMLogo;
